CLN: move create_activity_flag
