+++ docs/modules/analysis.rst
@@ -4,6 +4,11 @@ Analysis
 The analysis module provides a variety of functions to analyze the mobility
 datasets computed by trackintel. 
 
+Label
+=====
+
+.. autofunction:: trackintel.analysis.label.create_activity_flag
+
 Tracking Quality
 ================
