+++ trackintel/preprocessing/__init__.py
@@ -4,7 +4,6 @@ from .positionfixes import generate_triplegs
 from .filter import spatial_filter
 
 from .staypoints import generate_locations
-from .staypoints import create_activity_flag
 
 from .triplegs import smoothen_triplegs
 from .triplegs import generate_trips
